requirements.txt
ansible==4.7.0
ansible-core==2.11.6
asgiref==3.4.1
attrs==21.2.0
awsebcli==3.20.2
bcrypt==3.2.0
blessed==1.19.0
botocore==1.21.65
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
Django==3.2.8
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
future==0.16.0
graphviz==0.19.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==21.2
paramiko==2.8.1
pathspec==0.5.9
protobuf==3.17.3
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-lora==1.1.9
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
resolvelib==0.5.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.1.9
websocket-client==0.59.0
